rename and clean
